# Collection utils speedup for sized iterables