--- a/subprojects/base-services/src/test/groovy/org/gradle/util/CollectionUtilsTest.groovy
+++ b/subprojects/base-services/src/test/groovy/org/gradle/util/CollectionUtilsTest.groovy
@@ -126,8 +126,13 @@ class CollectionUtilsTest extends Specification {
         collect([1, 2, 3] as Object[], transformer { it * 2 }) == [2, 4, 6]
     }
 
-    def "collect iterable"() {
+    def "collect non-collection Iterable"() {
         expect:
+        def nonCollectionIterable = [iterator: { [1, 2, 3].iterator() }] as Iterable
+        collect(nonCollectionIterable, transformer { it * 2 }) == [2, 4, 6]
+    }
+
+    def "collect of collection iterable"() {
         collect([1, 2, 3] as Iterable, transformer { it * 2 }) == [2, 4, 6]
     }