@oehme 2019-04-27
Which real world performance issue does this solve? We really don't want to make the code any more complex if it's not for a good reason.