@paplorinc 2019-04-28
Hey @oehme, thanks for taking the time to review it.

I have rebased the commits to even smaller parts so that you can check them in their context: https://github.com/gradle/gradle/pull/9302/commits
* benchmark is in a separate commit, as proof of correctness (like a failing test before the fix);
* every commit that changed something needs proof, hence the benchmark comments were updated in the same commit;
* refactoring and logic change are in separate commits (like moving methods and changing their content);
* the last commit applies the logic to other places as well in a type-safe manner. If you still think it would be nicer to add the instanceof check inside the method (which I wanted to avoid, since I find it sneaky), I'll do it instead of the last commit.