@oehme 2019-04-29
I'm fine with merging the changes without the benchmark (I can guarantee you that no one will ever look at it again). However I suggest focussing on more impactful things afterwards. Optimizing things that are not an issue is not a good use of your valuable time.