@paplorinc 2019-04-29
@oehme, did this in the weekend, no task was yet assigned to me.
Could you please tell me what the reason is against the benchmarks, I treat them just like tests: if something changes later, we want to be able to reevaluate whether the previous choice is still the correct one.
They're in a special folder, it's easy to ignore them.