@oehme 2019-04-29
> Could you please tell me what the reason is against the benchmarks, I treat them just like tests: if something changes later, we want to be able to reevaluate whether the previous choice is still the correct one.

We evaluate changes based on our performance tests. It doesn't matter how something performs in an isolated micro-benchmark. What matters is the end-to-end performance of a Gradle build.