@paplorinc 2019-04-29
Thanks Stefan, removed the benchmarks, added instanceof and added new test that covers both branches.